refactor: merge `interface.py` module to `__init__.py`
